v2.0.8: Fix Admin Panel type casting error
BUG FIX: Fixed client-side exception in admin panel

Problem:
- Admin panel showing "client-side exception has occurred"
- PostgreSQL COUNT() and aggregate functions return bigint type
- bigint is serialized as string in JSON
- Frontend expects numeric types (number)
- TypeScript interface mismatch caused runtime error

Solution:
- Added explicit ::integer casts in SQL SELECT
- Cast fields: league_id, team_count, total_requests,
  unique_users, unique_managers
- Ensures JSON returns actual numbers, not strings

+++ src/app/api/admin/leagues/route.ts
@@ -27,12 +27,12 @@ export async function GET() {
         GROUP BY league_id
       )
       SELECT
-        ls.league_id,
+        ls.league_id::integer,
         'League ' || ls.league_id as league_name,
-        COALESCE(lt.team_count, 0) as team_count,
-        ls.total_requests,
-        ls.unique_users,
-        ls.unique_managers,
+        COALESCE(lt.team_count, 0)::integer as team_count,
+        ls.total_requests::integer,
+        ls.unique_users::integer,
+        ls.unique_managers::integer,
         ls.last_seen,
         ls.first_seen,
         ls.first_seen as created_at
